tripshorts/requirements.txt
moviepy==1.0.3
opencv-python==4.10.0.84
pydub==0.25.1
faster-whisper==1.0.1
numpy==1.26.4
pillow==10.4.0
rake-nltk==1.0.6
google-api-python-client==2.149.0
google-auth==2.35.0
tweepy==4.14.0
python-dotenv==1.0.1
